backend/requirements.txt
fastapi
uvicorn
python-multipart
spacy
sentence-transformers
google-generativeai
pdfplumber
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
